requirements.txt
streamlit>=1.33.0
python-docx==1.2.0
pydantic>=2.7.0
python-dotenv>=1.0.1
google-generativeai>=0.7.2
beautifulsoup4>=4.12.3
pypdf>=4.3.1
sentence-transformers>=3.0.1
chromadb>=0.5.5
